update Bridge: use chain specific gateway
